# spec/openapi.yaml
openapi: 3.1.0
info:
  title: StateScore API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /states:
    get:
      operationId: listStates
      parameters:
        - { name: region, in: query, schema: { type: string } }
      responses:
        '200': { description: State collection }
...
  /values:
    get:
      operationId: listMetricValues
      description: Returns all values when state_id is omitted, enabling bulk ranking loads.
      parameters:
        - { name: state_id, in: query, schema: { type: integer, format: int64 } }
        - { name: year, in: query, schema: { type: integer } }
      responses:
        '200': { description: Metric-value collection }
...